#89 Finished most of the UI for the ClassListFragment.
When starting no other tabs will appear.

Clicking on a class and send you to the tab with the chat, the title
will change to that of the class.

Can go back and select another tab and it will send you to the chat with
the title of the new selected class.

This does not reflect the actual logic, meaning you are not actually in
a new class chat (as of now).  All this is cosmetic changes used on the
front end.  Logic and server connection has not yet been implemented.